pyproject.toml
[project]
name = "oauth_shield"
version = "0.1.0"
description = "Protect static content behind oauth2 authentication"
authors = [
    {name = "Daniel Debonzi",email = "debonzi@gmail.com"}
]
# readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "requests-oauthlib (>=2.0.0,<3.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "sqlmodel (>=0.0.22,<0.0.23)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "jinja2 (>=3.1.5,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
